+++ src/cross_do.cpp
@@ -368,6 +368,12 @@ const int DO::nrec(const int gen_left, const int gen_right,
                    const bool is_x_chr, const bool is_female,
                    const Rcpp::IntegerVector& cross_info)
 {
+    #ifndef NDEBUG
+    if(!check_geno(gen_left, false, is_x_chr, is_female, cross_info) ||
+       !check_geno(gen_right, false, is_x_chr, is_female, cross_info))
+        throw std::range_error("genotype value not allowed");
+    #endif
+
     if(is_x_chr && gen_left > 36 && gen_right > 36) { // male X chromosome
         if(gen_left == gen_right) return(0);
         else return(1);