+++ src/cross_util.h
@@ -19,4 +19,12 @@ bool mpp_is_het(const int true_gen, const int n_alleles, const bool phase_known)
 const std::vector<std::string> mpp_geno_names(const std::vector<std::string> alleles,
                                               const bool is_x_chr);
 
+// For vector of founder indices (cross info), create the reverse index
+// with the reverse indices starting at 0
+// (2,3,1,4) -> (2,0,1,3)
+// (2,4,3,1) -> (3,0,2,1)
+// (7,8,3,5,4,1,6,2) -> (5,7,2,4,3,6,0,1)
+//
+Rcpp::IntegerVector reverse_index_founders(Rcpp::IntegerVector cross_info);
+
 #endif // CROSS_UTIL_H